requirements.txt
# Django والمتطلبات الأساسية
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# قاعدة البيانات
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# المصادقة والأمان
PyJWT==2.8.0
cryptography==41.0.7
django-oauth-toolkit==1.7.1

# Redis و Celery
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# معالجة الملفات والصور
Pillow==10.0.1
python-magic==0.4.27
reportlab==4.0.4

# HTTP requests للتكامل مع الخدمات الأخرى
requests==2.31.0
httpx==0.25.1

# معالجة البيانات
pandas==2.1.3
openpyxl==3.1.2

# التخزين السحابي
google-cloud-storage==2.10.0
boto3==1.29.7

# المراقبة والسجلات
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# الاختبارات
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.12.0

# الإنتاج
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8

# أدوات التطوير
black==23.10.1
flake8==6.1.0
isort==5.12.0

# معالجة التواريخ والوقت
python-dateutil==2.8.2
pytz==2023.3

# JSON والتحقق من البيانات
jsonschema==4.19.2
marshmallow==3.20.1

# ضغط الملفات
zipfile36==0.1.3

# إضافات Django مفيدة
django-extensions==3.2.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# WebSocket للإشعارات الفورية
channels==4.0.0
channels-redis==4.1.0

# معالجة المهام المجدولة
APScheduler==3.10.4

# التحقق من صحة البيانات
cerberus==1.3.5

# معالجة الـ CSV
csvkit==1.1.1

# أدوات إضافية
python-slugify==8.0.1
humanize==4.8.0
